+++ app/places_html_routes.py
@@ -1,7 +1,7 @@
-from flask import Blueprint, render_template, request, redirect
+from flask import Blueprint, render_template, request, redirect, flash
 from flask_jwt_extended import jwt_required
 from app.extensions import db
-from app.models import Place
+from app.models import Place, Show
 import uuid
 
 html_places_bp = Blueprint("html_places", __name__)
@@ -9,7 +9,16 @@ html_places_bp = Blueprint("html_places", __name__)
 @html_places_bp.get("/places")
 @jwt_required()
 def list_places():
-    places = Place.query.all()
+    search = request.args.get('search', '')
+    query = Place.query
+    
+    if search:
+        query = query.filter(
+            (Place.name.ilike(f'%{search}%')) | 
+            (Place.cidade.ilike(f'%{search}%'))
+        )
+    
+    places = query.all()
     return render_template("places_list.html", places=places)
 
 @html_places_bp.get("/places/new")
@@ -54,6 +63,14 @@ def edit_place(place_uuid):
 @jwt_required()
 def delete_place(place_uuid):
     place = Place.query.filter_by(uuid=place_uuid).first_or_404()
+    
+    # Verificar se o lugar está vinculado a algum show
+    show_linked = Show.query.filter_by(places_uuid=place_uuid).first()
+    
+    if show_linked:
+        flash("Não é possível apagar um lugar que está vinculado a um show!", "error")
+        return redirect("/places")
+    
     db.session.delete(place)
     db.session.commit()
     return redirect("/places")
